fix: resolve imports for testing
